requirements.txt
mysql-connector-python==8.0.32
PyQt6==6.4.2